✅ test: 테스트 디렉토리 구조 server와 맞게 변경
