+++ feed-crawler/package.json
@@ -40,9 +40,9 @@
     "build": "tsc --project tsconfig.build.json",
     "start": "node dist/src/main.js",
     "start:dev": "ts-node src/main.ts",
-    "test:unit": "cross-env NODE_ENV=test jest --config test/jest-unit.json",
-    "test:unit:cov": "cross-env NODE_ENV=test jest --coverage --config test/jest-unit.json",
-    "test:e2e": "cross-env NODE_ENV=test jest --config test/jest-e2e.json",
-    "test:e2e:cov": "cross-env NODE_ENV=test jest --coverage --config test/jest-e2e.json"
+    "test:unit": "cross-env NODE_ENV=test jest --config test/config/unit/jest/jest-unit.json",
+    "test:unit:cov": "cross-env NODE_ENV=test jest --coverage --config test/config/unit/jest/jest-unit.json",
+    "test:e2e": "cross-env NODE_ENV=test jest --config test/config/e2e/jest/jest-e2e.json",
+    "test:e2e:cov": "cross-env NODE_ENV=test jest --coverage --config test/config/e2e/jest/jest-e2e.json"
   }
 }
