+++ feed-crawler/test/unit/feed-crawler.spec.ts
@@ -56,6 +56,7 @@ describe('FeedCrawler', () => {
   ];
 
   beforeEach(() => {
+    jest.clearAllMocks();
     mockFeedRepository = {
       deleteRecentFeed: jest.fn(),
       insertFeeds: jest.fn(),
